tesla-protocol: address review feedback
- Validate peer ECDH public keys are on the curve (mbedtls_ecp_check_pubkey)
  before the scalar multiply; the reference implementation rejects off-curve
  peers and the old code only caught unparseable input. Add an off-curve
  (well-formed, 0x04-prefixed) negative test.
- Widen the metadata builders' flags parameter to uint32_t to match the wire
  field (RoutableMessage.flags); the reference serializes the full 32-bit
  value, so any future flag bit >= 8 must survive into the AAD.
- Move the two-part 'M || payload' HMAC behind a tesla_hmac_sha256_2 helper
  in crypto.h so session.c no longer reaches into mbedTLS directly.
- CI: retry + SHA-256-verify the mbedTLS tarball, cache the host build prefix.
- Test/doc cleanups: drop dead vector constants, pass truncate_to_17=false in
  the Infotainment round trip (AES-GCM tag is already 16 bytes), document the
  duplicate-tag strictness vs the reference, and fix dangling doc links.

+++ components/tesla-protocol/session.h
@@ -53,8 +53,11 @@
 #define TESLA_METADATA_MAX 128
 
 // Tag-length-value metadata serializer. Tags must be added in strictly
-// ascending order (mirrors the reference implementation); a value longer than
-// 255 bytes is rejected because the TLV length field is a single byte.
+// ascending order. This is deliberately stricter than the reference
+// implementation, which permits duplicate tags: the serialization stays
+// injective (no two metadata sets collide), which the protocol relies on. A
+// value longer than 255 bytes is rejected because the TLV length field is a
+// single byte.
 typedef struct {
     uint8_t buf[TESLA_METADATA_MAX];
     size_t  len;
@@ -97,19 +100,19 @@ bool tesla_verify_session_info(const uint8_t k[TESLA_SHARED_KEY_LEN],
 
 // Serialized metadata for an outgoing AES-GCM command. The AAD for encryption
 // is SHA256(serialized). flags is only encoded when non-zero (backwards
-// compatibility rule in protocol.md).
+// compatibility rule in protocol.md); it is a full 32-bit value on the wire.
 int tesla_build_request_metadata(tesla_metadata_t *m, uint8_t domain,
                                  const uint8_t *vin, size_t vin_len,
                                  const uint8_t epoch[TESLA_EPOCH_LEN],
                                  uint32_t expires_at, uint32_t counter,
-                                 uint8_t flags);
+                                 uint32_t flags);
 
 // Serialized metadata for decrypting an encrypted response (AAD = SHA256(serialized)).
 // Per protocol.md the Flags field is always included here (unlike requests) and
 // the fault is always encoded (0 when none).
 int tesla_build_response_metadata(tesla_metadata_t *m, uint8_t domain,
                                   const uint8_t *vin, size_t vin_len,
-                                  uint32_t counter, uint8_t flags,
+                                  uint32_t counter, uint32_t flags,
                                   const uint8_t *request_hash, size_t request_hash_len,
                                   uint32_t fault);
